requirements.txt
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jsonapi==2.7.0
inflection==0.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
